fix(home): eliminate duplicate API requests on homepage
